send_rangefinder_vision.py: cast time to int so mavlink is able to pack it properly

+++ developers/pymavlink/send_rangefinder_vision.py
@@ -7,7 +7,7 @@ def wait_conn(master):
     msg = None
     while not msg:
         master.mav.ping_send(
-            time.time(), # Unix time
+            int(time.time()), # Unix time
             0, # Ping number
             0, # Request ping of all systems
             0 # Request ping of all components
@@ -43,7 +43,7 @@ tstart = time.time()
 while True:
     time.sleep(0.5)
     master.mav.distance_sensor_send(
-        (time.time() - tstart) * 1000,
+        int((time.time() - tstart) * 1000),
         min,
         max,
         distance,
